Factored out importWallet
