+++ app/api/aave/route.tsx
@@ -2,7 +2,7 @@ import { NextResponse } from 'next/server';
 import { baseSepolia } from 'viem/chains';
 import aaveAbi from './aave_v3_abi.json';
 import usdcAbi from './usdc_abi.json';
-import '@/lib/coinbase';
+import { importWallet } from '@/lib/coinbase';
 
 import {
     createPublicClient,
@@ -313,51 +313,4 @@ export async function POST(request: Request) {
     }
 
     return NextResponse.json({ error: 'Invalid action' }, { status: 400 });
-}
-
-// importWallet imports the CDP wallet from the environment variables.
-async function importWallet(): Promise<Wallet> {
-    const { WALLET_DATA } = process.env;
-
-    try {
-        // Parse the wallet data
-        const seedData = JSON.parse(WALLET_DATA || "{}");
-
-        if (Object.keys(seedData).length === 0) {
-            // Create a new wallet if WALLET_DATA is empty
-            const newWallet = await Wallet.create();
-
-            // Fund the wallet with USDC
-            let faucetTransaction = await newWallet.faucet(Coinbase.assets.Usdc);
-            console.log(`Faucet transaction for USDC: ${faucetTransaction}`);
-
-            // Fund the wallet with ETH.
-            faucetTransaction = await newWallet.faucet();
-            console.log(`Faucet transaction for ETH: ${faucetTransaction}`);
-
-            let exportData = await newWallet.export()
-
-            // Create and assign the new WALLET_DATA
-            const newWalletData = JSON.stringify({ [exportData['walletId'] as string]: { 'seed': exportData['seed'] as string } });
-
-            console.log(`Created new wallet: ${exportData['walletId']}`)
-
-            process.env.WALLET_DATA = newWalletData;
-
-            return newWallet;
-        }
-
-        // Get the wallet id
-        const walletId = Object.keys(seedData)[0];
-
-
-        // Get the seed of the wallet
-        const seed = seedData[walletId]?.seed;
-
-        // Import the wallet
-        return Wallet.import({ seed, walletId });
-    } catch (e) {
-        console.log('Failed to import wallet:', e);
-        throw e;
-    }
 }
